requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.5.0
python-dateutil==2.9.0.post0
pytz==2023.3
pyyaml>=6.0

# Web framework and interface
streamlit>=1.22.0
streamlit-autorefresh==0.0.2
plotly>=5.13.0
matplotlib==3.10.1
watchdog==3.0.0

# Data handling and processing
ccxt==4.2.14
pandas-ta==0.3.14b0
ta>=0.10.2
scikit-learn>=1.0.2
scipy==1.15.2
joblib==1.4.2

# Machine learning
mlflow>=2.3.0
tensorflow==2.12.0
xgboost==1.7.6
prophet==1.1.4
optuna==3.3.0
statsmodels==0.14.0

# Database
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
alembic==1.12.0

# Utilities
requests==2.32.3
python-dotenv>=0.19.0
tqdm==4.66.4
appdirs==1.4.4
tabulate==0.9.0
colorlog==6.7.0
schedule==1.2.0
cryptography==41.0.3

# Notifications
python-telegram-bot==20.4
aiosmtplib==2.0.2

# Testing
pytest>=7.0.0
pytest-mock==3.11.1

# Documentation
sphinx==7.2.5
sphinx-rtd-theme==1.3.0

# Development tools
black>=22.3.0
isort==5.12.0
flake8>=4.0.1
mypy>=0.950

# Additional dependencies
seaborn==0.13.2
logging_setup==0.0.9
pyarrow==19.0.1

# New dependencies
python-binance>=1.0.16
